requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==1.2.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.1
multidict==6.1.0
numpy==2.1.2
pandas==2.2.3
parsimonious==0.10.0
propcache==0.2.0
psycopg2-binary==2.9.9
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
toolz==1.0.0
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
web3==7.3.0
websockets==13.1
yarl==1.14.0
